fix: use wrapt_timeout_decorator, instead of stopit  (#2938)
<!--Add a description of your PR here-->

There are some issues of the solver stalling (or not behaving as
expected) when running very large DAGs (e.g. #871, #1003, #1374, #1620,
#2354).

To minimize those issues, I was wondering if it would make sense for the
solver/scheduler to evaluate just a subset (randomly sampled) of the
jobs waiting to run.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced job scheduling with improved timeout handling using
asynchronous programming for better performance and responsiveness.
- **Chores**
- Updated project dependencies, replacing `stopit` with
`timeout_decorator` to support modern asynchronous handling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>