chore: For bundled assets, reference the particular version’s license
Referencing the license file from the release tag corresponding to the
bundled release of each asset makes it less likely that the checksum
will eventually stop matching due to later upstream license changes –
perhaps as simple as a new copyright date – and makes errors in the
included license less likely.

Since the copyright statement in the react license changed from
"Copyright (c) Facebook, Inc. and its affiliates." to "Copyright (c)
Meta Platforms, Inc. and affiliates." since the bundled version of react
was released – and this commit uses the older version that corresponds
to the bundled library – the checksum of the react license file changes.

Similarly, in the vega license, the copyright date range changed from
2015-2021 to 2015-2023 after the bundled version was released, so using
the correct license file for the bundled version changes the checksum.