chore: For assets bundled at a particular version, reference the version’s license (#3098)
Referencing the license file from the release tag corresponding to the
bundled release of each asset makes it less likely that the checksum
will eventually start to match due to later upstream license changes –
perhaps as simple as a new copyright date – and makes errors in the
included licenses less likely.

Since the copyright statement in the react license changed from
"Copyright (c) Facebook, Inc. and its affiliates." to "Copyright (c)
Meta Platforms, Inc. and affiliates." since the bundled version of react
was released – and this commit uses the older version that corresponds
to the bundled library – the checksum of the react license file changes.

Similarly, in the vega license, the copyright date range changed from
2015-2021 to 2015-2023 after the bundled version was released, so using
the correct license file for the bundled version changes the checksum.

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case. **N/A – this change is not
really testable – except that building the package confirms the
checksums match.**
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake). **N/A – no
documentation update is believed to be necessary.**


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Updated asset URLs to point to specific tagged versions for improved
stability and reliability.

- **Bug Fixes**
- Corrected SHA256 checksums for various assets to ensure integrity and
security.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->