fix: rely on conda >24.7.1 for conda env deployment and deprecate mamba support (#3121)
Rationale:
Since a while, conda uses internally the libmamba solver (i.e. the same
thing as mamba). Mamba development itself is basically over, as the devs
concentrate on rattler-based solving. The latter is the future for
Snakemake as well, by natively interfacing with rattler via py-rattler.
However, until then, the recent quite disruptive changes in mamba 2.0
(including CLI and unexpected behavior changes) suggest that it is more
robust for us at the moment to move back to conda, and simply require a
recent conda release, such that it is as fast as mamba.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Updated the default package manager from `mamba` to `conda` across
various components, simplifying environment setup.
- Introduced a new function to enhance environment setup flag
management.

- **Bug Fixes**
- Enhanced error handling and messaging for conda environment
initialization.

- **Chores**
- Removed unnecessary dependencies, including `mamba`, and updated the
testing environment configuration for better compatibility.
- Updated Docker images and commands for environment setup to reflect
the transition to `conda`.
- Added the `nodefaults` channel to configuration files for improved
package management.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->