fix: remote pre-command not installing all required storage plugins (#3116)
<!--Add a description of your PR here-->
Related Issues: 
#2501 

https://github.com/snakemake/snakemake-executor-plugin-kubernetes/issues/19

When the job spawner prepares the `precommand` for remote execution, it
only registers the `default storage provider` which fails to acknowledge
the presence of a storage provider within the snakefile.

This PR should now register all storage providers that `workflow` has
recognized.

NOTE: unsure how to I could add a test case for this? 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced package installation logic for storage providers, allowing
multiple packages to be installed simultaneously when the auto-deploy
feature is enabled.

- **Bug Fixes**
- Improved handling of duplicate package names in the installation
process.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->