: getting rid of logger wrapper class
The logger wrapper class (seemingly) largely existed to enable the custom log methods (job_info, progress, etc). However this was quite confusing, especially the usage of the logRecord msg as a dictionary to pass along conextual information.

 In this commit, the logger class is no longer used. Direct calls to log.info/warn/debug use the `extra` argument to pass a dict of contextual information (jobid, etc). Then the default log formatter/filter can use this info to format and filter logs respectively. This approach provides a much more explicit route of log records and allows potential log plugins to handle logs however they see fit.

I've put some small functions to setup the logger and such, however this is largely to get things running and I will put more effort into shoring this up if we think this is a good approach.