fix: Cannot change working directory when R script hosted on GitHub with source (#3164)
<!--Add a description of your PR here-->

When running a workflow with a Snakefile using the `github` function
(see below) which references an R script which sources a file an error
like, "Cannot change the working directory" appears.

This changes the source logic to only change the working directory if
the script is not a URL. I also changed the code to use
`on.exit(setwd(old_wd), add = TRUE)` instead of relying on the final
`setwd(wd)` to be a little safer in case there's any errors later. That
way the user's working directory is not changed if
`snakemake@source(...)` is in a try-catch-except block.

#### Snakefile
```
module module_name:
    snakefile:
        github("GITHUB_REPO", path="workflow/Snakefile", tag="v0.1.5")
    config:
        config
```

#### R script in module_name
```
...
# Fails on next line
snakemake@source("file_hosted_on_github")
...
```

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Improved handling of script directories in R scripts, ensuring
reliability when sourcing scripts from various locations.
- Enhanced logic for generating preambles in R scripts to reflect the
new directory handling.

- **Bug Fixes**
- Resolved issues related to changing the working directory when the
script directory is a URL.

- **Documentation**
	- Minor adjustments made for clarity in comments and formatting.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: rschauner <rschauner@fastmail.com>