fix: correct formatting of R preamble (#2425)
### Description

Currently parameters given to the R script preamble do not work for some
(uncommon) cases:

- special float values (NaN, inf, -inf)
- complex numbers
- integers (type preserved)
- heterogeneous lists

These cases are corrected in this PR.

Heterogeneous lists are now supported by creating an R `list` instead of
a R `vector`, if the Python `list` contents are not from the same type.
Therefore, this change may break some workflows, if they rely on the R
script input being a vector or the implicitly converted types (e.g.
before this PR `TRUE` became `"TRUE"` if there is another string in the
list). In my opinion, I would regard the implicit type conversion a bug,
hence my change proposal.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>