fix: Improved handling of missing output files in group job postprocessing, accounting for temporary files. (#1765)
When I run group jobs on a cluster, I'll often write temporary files to
the node-specific local scratch folder. I'll label them `temp(...)` and
the sub-snakemake instance will dutifully delete the files when no
longer needed. Even without that deletion, the files would disappear
with the node local scratch as soon as the group job finishes.

The problem occurs when the parent snakemake instance re-assumes
control. It checks for all the outputs from the group job, finds the
temp files missing, and exits with an error. But these temp files are no
longer needed anywhere in the workflow, and indeed, Snakemake's
immediate next step if it were to find the temp files would be to delete
them.

This PR filters the group job outputs before the existence checks,
removing any that are:
- Not needed anywhere else in the workflow
- Not target files

In particular, only files that would otherwise be immediately removed
are excluded from the existence check (at least, that's the idea).

The idea is to remove a small frustration without causing any breaking
or generally noticeable changes. I've no test yet, but I'll write one if
the concept here is approved.


### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced output checking mechanism to selectively ignore specific
output files.
- Improved handling of missing output files in job postprocessing,
accounting for temporary files.

- **Bug Fixes**
- Refined logic for determining if files are needed by pending jobs,
improving accuracy in output management.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>