fix: Revert cleaning of env vars in apptainer (#3285)
<!--Add a description of your PR here-->

This PR reverts the change introduced by
76d53290a003891c5ee41f81e8eb4821c406255d which, when using
Apptainer/Singularity, forces cleaning of system environment variables
going into the container. This is neither a fixed nor default behavior
of Apptainer itself since it can be desirable to not fully isolate from
the system. Hence Snakemake shouldn't dictate it. With the change in
76d53290a003891c5ee41f81e8eb4821c406255d, users do not have any chance
to deactivate the behavior. On the other hand, the `--cleanenv` can
easily be added through `--aptainer-args`.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
- Updated container execution behavior by removing the `--cleanenv`
option, allowing host environment variables to be passed through to
container tasks for a more consistent execution context.
  
- **Documentation**
- Enhanced documentation on Snakemake workflow structure, adding
requirements for a "Code of Conduct," "Contribution instructions," and a
"Test directory."
- Clarified the importance of maintaining a `.tests` directory and
specified subdirectories for integration and unit tests.
- Updated repository structure requirements to include an SVG-formatted
rule graph and technical documentation file.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->