Add license files for dependencies bundled in vega-lite.js
These are the NPM packages that are included in the vega-lite.js bundle
via dependencies. An effort has been made to verify that every license
file added in this commit actually corresponds to some code present in
vega-lite.js.

Note that no license files needed to be added for software
copied/derived/adapted into vega-lite.js; hashlru was bundled via
vega-util, but this was already covered for vega.js.