Add license files for dependencies bundled in vega-embed.js
These are the NPM packages that are included in the vega-embed.js bundle
via dependencies. An effort has been made to verify that every license
file added in this commit actually corresponds to some code present in
vega-embed.js.

Note that no license files needed to be added for software
copied/derived/adapted into vega-embed.js; everything is already covered
via the dependencies for vega-lite.js or vega.js.