refactor: Remove datrie dependency (#3176)
This PR proposes to replace the datrie dependency, which is causing
installation problems now (due to native compilation failing on some
newer platforms).

The logic which used tries before has been extracted into a class
`PrefixLookup` which allows storing (string) key and (arbitrary) value
pairs allowing multiple values for the same key. It provides
functionality to obtain all values for which their corresponding key is
a prefix of the specified query.

This lookup is performed in `O(m log(n))` (thanks again @tbooth) where m
is the number of matching results, and n is the length of the entries
list being searched. When constructing the `PrefixLookup` inedx
initially the cost depends on the number of unique keys, as they will be
sorted.

The PR also adds some unit tests for this functionality.

Fixes #3105 and
Fixes #2970 

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

## Release Notes

- **New Features**
	- Introduced a new `PrefixLookup` class for efficient prefix matching.

- **Bug Fixes**
- Removed the `datrie` dependency across various configuration files,
simplifying installation requirements.

- **Documentation**
- Updated the FAQ to remove redundant phrases related to compiler
errors.

- **Tests**
- Introduced comprehensive unit tests for the `OutputIndex` class,
covering various matching scenarios and edge cases.
- Added unit tests for the new `PrefixLookup` class to validate its
functionality.
- Updated tests to ensure robustness in matching functionalities under
various conditions.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Tim Booth <tim.booth@ed.ac.uk>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>