fix: update syntax, folding, and sections vim files (#3340)
Updates the vim files with to align with added directives, statements,
and functions. Closes #1114

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced code folding in Snakemake scripts for clearer and more
dynamic organization.
- Improved section navigation, making it easier to jump between key
parts of your Snakemake files.
- Updated syntax highlighting with support for new Snakemake constructs
and parameters, offering a modernized editing experience.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->