feat: shell function calls inside of the 'run' directive now use conda, container, or envmodules specifications (#2289)
### Description

Rules that combine python and shell code can be quite useful, e.g.:
```
  run:
       x = parse_input_file(input)
       shell('command --param {x}')
```

Conda integration for the shell commands in such a rule is currently
only supported like this:
```           
       shell('command --param {x}', conda_env='named_preexisting_environment')
```
Named pre-existing conda environments make a workflow however more
difficult to distribute and could lead to reproducibility issues.

Better would be to use conda specification files (e.g.
'envs/my_env.yaml'). This is however not supported by 'shell' function
calls. Implementing it would not be straightforward, because these
shell-calls are only executed during the processing of a rule, and
snakemake therefore would not know at startup which conda environments
would be needed. Creating the environments as-needed during rule
execution would lead to issues with rules that are executed in parallel.

This issue has bee noted before #163, #1619. One possible solution is to
put all code within e.g. a separate python script which in turn executes
'command'. Then everything can run within the conda environment. This
however requires quite a bit of boilerplate code to substitute what
snakemake normally does (e.g. argument parsing, formatting of the shell
command based on params/resources/input/output files, subprocess
machinery). Another alternative is to implement this with a bash script
and use the 'shell' directive, but this is usually not as readable as a
python solution, particularly if one needs to do more complex
operations, e.g. parse an input file.

This pull request therefore enables the following rule structure: 
```
  conda: 'envs/my_env.yaml'
  run:
       x = parse_input_file(input)
       shell('command --param {x}')
```
This will perform 'command --param {x}' within the specified conda
environment. Note that it will not perform the whole run block within
the conda environment, but only shell function calls that are executed
from the run block. A note specifying this is added to the
documentation. For symmetry, it makes this change also for the
'container' directive.

No large code changes are required, as the functionality is already
present, only a check has to be removed.

One limitation of this solution is that each shell call executed from
the run script will be performed in the same conda environment (it is
though possible to execute without the conda environment using
`shell('command', conda_env=None)`).
A future implementation could make this more flexible by enabling
something like this:
```
conda:
      env1='envs/env1.yaml',
      env2='envs/env2.yaml'
run:
    if <condition>: 
         shell('command1 {input}', conda_env=conda_env.env1)
    else:
         shell('command2 {input}', conda_env=conda_env.env2)
```



### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Documentation**
- Revised guidance now allows using conda environments and container
setups with an additional workflow directive, enhancing clarity on
environment execution.
  
- **New Features**
- Expanded support for environment and container configurations
increases workflow flexibility.
- Introduced a new rule for executing a Python script within a conda
environment.
  
- **Tests**
- Added new test cases to validate the execution of conda environments
and updated existing tests to reflect changes in script paths and
outputs.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>