fix: Do not perform storage object cleanup with --keep-storage-local-copies set (#3358)
This PR fixes the `--keep-storage-local-copies` functionality. Prior to
these changes, the option would retain the dirnames of the output files,
but the files themselves would be deleted.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Updated the help text for the "--keep-storage-local-copies" option to
clarify that local copies are kept for both remote input and output
files.
  
- **New Features**
- Enhanced workflow storage management to conditionally preserve storage
objects when local copies are intended to be retained.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: Tadas Bareikis <tadas@vugene.com>