fix: Prevent execution of conda in apptainer when not explicitly requested in software deployment method (#3388)
This is a quick fix for the issue described in #3202.

From the original issue:
> Snakemake crashes with a 'conda info --json' returned non-zero exit
status 127 when a rule contains both conda and container directives.

This behaviour was caused by the retrieval in `_software_stack_hash()`
within the `Persistence` class. If a snakemake rule specified both a
container and a conda file, this method would always try to get a hash
value of the conda file, resulting in an instantiation of the `Conda`
class using the container as the base image (ad-hoc combination). This
happened even when conda was not specified with `--sdm`. This also
happened in `get_log_error_info' of `jobs.py'. This PR fixes this issue
and ensures that rules can be executed wither with conda, apptainer or
the ad-hoc combination.

### QC

* [ x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Enhanced handling of environment details to prevent errors when
specific deployment modes aren’t active.
  - Improved logic for updating environment-related configurations.

- **New Features**
- Introduced a workflow that supports execution in both containerized
and Conda environments.
- Added a dependency configuration file for consistent package
management.
  - New output file created with expected results from the workflow.
- Expanded validation capabilities for data structures, including
support for both Pandas and Polars.

- **Tests**
- Implemented new tests to verify proper execution across varying
deployment methods.
  - Enhanced validation tests for sample data against specified schemas.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->