fix: print filenames with quotes around them in RuleException (#3269)
Currently RuleException formats filenames without quotes, so errors look
like this:

```
TypeError in file /.../workflow/rules/detection.smk, line 197:
'str' object is not callable
  File "/.../workflow/rules/detection.smk", line 192, in __rule_fit_gaussian
  File "/.../workflow/rules/detection.smk", line 197, in <listcomp>
```

This is a minor issue, but for example when using ctrl+click on a
filename in VSCode, instead of opening the "detection.smk" it tries to
open "detection.smk," which is annoying. This PR makes it so the first
filename is also printed with quotes, in line with Python's default
formatting of filenames in stacktraces.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Improved error message formatting to clearly indicate filenames in
exception messages by enclosing them in double quotes.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>