feat!: Logging refactor & add LoggerPluginInterface (#3107)
Followup to #2830.

This PR attempts to make Snakemake's logging more modular to make room
for a potential logging plugin interface, or other logging/monitoring
additions, as well as make logging in Snakemake more standardized.

Currently, most of logging logic is deep within `logging.Logger`, mainly
in `text_handler()` with direct calls to `logger.info/warn/etc`. In this
PR, this logic has been refactored to more comply with the standard
`logging` practices. To do so, three new classes were added:
1. `ColorizingTextHandler` manages adding color to log records and
emitting to stream.
2. `DefaultFormatter` formats log records to be printed. Much of the
logic from `text_handler()` is here now.
3. `DefaultFilter` filters log records from being emitted based on level
and quietness settings.

Together, these classes should output identical logs to the current
logging implementation. I've tested this briefly, but further testing
would be good.

With these changes, it is much easier to add new `logHandlers` via
`logger.addHandler`. New handlers can decide how to format and filter
logs themselves, an important characteristic missing in the current
implementation that motivated this refactoring
(https://github.com/snakemake/snakemake/pull/2830#issuecomment-2092731491).

This would be a breaking change for `--log-handler-script`, as user
provided log handlers will need to subclass `logging.logHandler`.
Furthermore, logic for adding the user handler (and formatter/filter)
will need to be added.

As for the logging plug interface
(https://github.com/snakemake/snakemake/pull/2830#issuecomment-2091510808),
I'm not sure that is the route to take, but would appreciate thoughts on
this. First, this functionality is already largely provided by
`--log-handler-script` and adding a logging plugin would be redundant.
Second, I think that development of such plugins would likely be
minimal, and coordinated effort in this repo to maintain a good set of
logging options would be preferable. For example:
- A standard HTTP log handler send emit log records to a user supplied
address on which a 3rd party monitoring app is listening.
- A handler/formatter that uses
[rich](https://github.com/Textualize/rich) to make a
prettier/interactive console on which logs are written to.

I think both of these additions would benefit being standardized to
Snakemake, as well as default options to users.

Additionally I think it would be worthwhile to unify logging in
Snakemake. Currently, the use of the wrapper `logging.Logger` class and
`setup_logger()` function being called in multiple places makes it
difficult to work with the logger and is likely the source of some
issues (#3044, #2974, #2882, #2797, #2727). Unifying the logger and its
configuration as recommended by the [logging
docs](https://docs.python.org/3/howto/logging-cookbook.html#a-cli-application-starter-template)
would make it easier to work the the logger. Another option would be to
use something like
[loguru](https://loguru.readthedocs.io/en/stable/index.html). Whichever
option, I think simplifying and unifying the logger should be the next
step of this PR.

One main question I have is how much we would like to preserve the
current stderr output formatting. Most of logging logic is in formatting
rule info and such. Moving forward we can now try different formatting.
Personally, I find the current detailed rule output helpful for sanity
checks when developing and doing a dry run. However, when actually
running the workflow (especially a large one) the terminal is flooded
and errors can quickly be lost which requires searching the log file. Of
course this can be alleviated with --quiet.

Another question is regarding the two calls of setup_logger() brought up
in #2797. I still don't understand why these are necessary and
understanding this would be helpful in unifying this.

I'm happy to keep pushing this forward but would appreciate your
thoughts before continuing @johanneskoester! I


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new CLI option (--logger) for configuring custom loggers.
- Enhanced structured logging delivers clearer runtime and resource
insights, including rule graph logging.
	- Added a new method for logging rule graph information.
	- Added a new enumeration value `REASON` to the `Quietness` class.

- **Refactor**
- Overhauled the logging architecture across execution, scheduling, and
workflow components for improved traceability and error handling.

- **Chores**
- Updated dependencies and setup configurations to include new logger
plugins.
- Adjusted test configurations to capture all output, ensuring more
comprehensive test diagnostics.

- **Bug Fixes**
- Resolved minor command and formatting inconsistencies for smoother
operations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>